refactor(components): domínio single-módulo sai da global → Pages/<Mod>/_components
Camada 4 (Módulo · ADR UI-0013) não vive mais em Components/ global quando
só 1 módulo consome — alinha com convenção _components/ já vigente (Sells,
CaixaUnificada, ServiceOrders):

- MOVE Components/clientes/{ActiveChip,Avatar,KpiStripClickable,Pills} → Pages/Cliente/_components/
- MOVE Components/ConsultaOs/{OsLookupForm,OsPipeline,OsResultCard,OsStageBadge} → Pages/ConsultaOs/_components/
- MOVE Components/jana/AssistantUiChat → Pages/Jana/_components/

Ficam na global (justificado): board (cross-módulo OficinaAuto+ProjectMgmt),
cockpit (Shell), Site (surface pública + SiteLayout), NfeBrasil (domínio
fiscal consumido por Sells).

Refs re-path: imports (3 páginas) · Pest GUARDs 1/2/3/13 · eslint-baseline
(5 keys + absorve 6 no-restricted-syntax pré-existentes que entraram no
escopo DS rules Pages/**) · layout-primitives-baseline (6 keys, counts
idênticos) · charter Cliente · docs vivos Crm/REUSE_MAPPING · CHANGELOG DS
0.6.9 com entrada MOVE.

Gates locais: pageheader:guard ✅ 104=104 · layout:check ✅ · reuse:gate ✅ ·
lint:baseline:check ✅ Δ-23 · a11y:check ✅ 294=294 · conformance:check ✅ ·
zero TS2307 novos (423 erros tsc pré-existentes inalterados).

Zero mudança visual — refactor de path puro.

Refs: CYCLE-08 · Constituição UI v2 ADR UI-0013 camada 4
