+++ resources/js/Pages/ConsultaOs/Index.tsx
@@ -7,8 +7,8 @@ import { ArrowLeft, AlertCircle, FileSearch } from 'lucide-react'
 import { Button } from '@/Components/ui/button'
 import { Card, CardContent, CardHeader } from '@/Components/ui/card'
 import { Alert, AlertDescription } from '@/Components/ui/alert'
-import { OsLookupForm } from '@/Components/ConsultaOs/OsLookupForm'
-import { OsResultCard } from '@/Components/ConsultaOs/OsResultCard'
+import { OsLookupForm } from './_components/OsLookupForm'
+import { OsResultCard } from './_components/OsResultCard'
 import type { OrdemServico } from '@/Types/os'
 
 type Estado = 'busca' | 'resultado' | 'nao-encontrado'
