+++ resources/js/Pages/ConsultaOs/_components/OsPipeline.tsx
