+++ resources/js/Pages/ConsultaOs/_components/OsResultCard.tsx
