+++ resources/js/Pages/ConsultaOs/_components/OsStageBadge.tsx
