+++ resources/js/Pages/Jana/_components/AssistantUiChat.tsx
