+++ scripts/layout-primitives-baseline.json
@@ -14,13 +14,13 @@
   "files": {
     "resources/js/Components/board/BoardColumn.tsx": 2,
     "resources/js/Components/board/TaskCard.tsx": 2,
-    "resources/js/Components/clientes/Avatar.tsx": 1,
-    "resources/js/Components/clientes/KpiStripClickable.tsx": 2,
+    "resources/js/Pages/Cliente/_components/Avatar.tsx": 1,
+    "resources/js/Pages/Cliente/_components/KpiStripClickable.tsx": 2,
     "resources/js/Components/CommandPalette.tsx": 1,
-    "resources/js/Components/ConsultaOs/OsLookupForm.tsx": 5,
-    "resources/js/Components/ConsultaOs/OsPipeline.tsx": 2,
-    "resources/js/Components/ConsultaOs/OsResultCard.tsx": 3,
-    "resources/js/Components/jana/AssistantUiChat.tsx": 9,
+    "resources/js/Pages/ConsultaOs/_components/OsLookupForm.tsx": 5,
+    "resources/js/Pages/ConsultaOs/_components/OsPipeline.tsx": 2,
+    "resources/js/Pages/ConsultaOs/_components/OsResultCard.tsx": 3,
+    "resources/js/Pages/Jana/_components/AssistantUiChat.tsx": 9,
     "resources/js/Components/MentionInput.tsx": 1,
     "resources/js/Components/PageHeader/PageHeader.tsx": 2,
     "resources/js/Components/shared/BulkActionBar.tsx": 1,
